+++ modules/generic/src/main/scala/pureconfig/generic/CoproductHint.scala
@@ -3,6 +3,7 @@ package pureconfig.generic
 import com.typesafe.config.{ ConfigFactory, ConfigObject, ConfigValue, ConfigValueType }
 import pureconfig._
 import pureconfig.error._
+import pureconfig.generic.error.{ NoValidCoproductChoiceFound, UnexpectedValueForFieldCoproductHint }
 import pureconfig.syntax._
 
 /**
@@ -46,6 +47,16 @@ trait CoproductHint[T] {
    * @return `true` if the next class or coproduct option should be tried, `false` otherwise.
    */
   def tryNextOnFail(name: String): Boolean
+
+  /**
+   * Returns a non empty list of failures scoped into the context of a `ConfigCursor`, representing the failure to read
+   * an option from the config value.
+   *
+   * @param cur a `ConfigCursor` at the sealed family option
+   * @return a non empty list of reader failures.
+   */
+  def noOptionFound(cur: ConfigCursor): ConfigReaderFailures =
+    ConfigReaderFailures(cur.failureFor(NoValidCoproductChoiceFound(cur.value)))
 }
 
 /**
@@ -54,7 +65,7 @@ trait CoproductHint[T] {
  * This hint will cause derived `ConfigConvert` instance to fail to convert configs to objects if the object has a
  * field with the same name as the disambiguation key.
  *
- * By default, the field value written is the class or coproduct option name converted to lower case. This mapping can
+ * By default, the field value written is the class or coproduct option name converted to kebab case. This mapping can
  * be changed by overriding the method `fieldValue` of this class.
  */
 class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
@@ -65,7 +76,7 @@ class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
    * @param name the name of the class or coproduct option
    * @return the field value associated with the given class or coproduct option name.
    */
-  protected def fieldValue(name: String): String = name.toLowerCase
+  protected def fieldValue(name: String): String = FieldCoproductHint.defaultMapping(name)
 
   def from(cur: ConfigCursor, name: String): ConfigReader.Result[Option[ConfigCursor]] = {
     for {
@@ -90,6 +101,15 @@ class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
   }
 
   def tryNextOnFail(name: String) = false
+
+  override def noOptionFound(cur: ConfigCursor): ConfigReaderFailures =
+    cur.fluent.at(key).cursor.fold(
+      identity,
+      cur => ConfigReaderFailures(cur.failureFor(UnexpectedValueForFieldCoproductHint(cur.value))))
+}
+
+object FieldCoproductHint {
+  val defaultMapping: String => String = ConfigFieldMapping(PascalCase, KebabCase)
 }
 
 /**
@@ -98,6 +118,7 @@ class FieldCoproductHint[T](key: String) extends CoproductHint[T] {
  *
  * @tparam T the type of the coproduct or sealed family for which this hint applies
  */
+@deprecated("Use `pureconfig.generic.semiauto.deriveEnumerationReader[T]`, `pureconfig.generic.semiauto.deriveEnumerationWriter[T]` and `pureconfig.generic.semiauto.deriveEnumerationConvert[T]` instead", "0.10.3")
 class EnumCoproductHint[T] extends CoproductHint[T] {
 
   /**