+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+### 0.10.3 (unreleased)
+
+- Breaking changes
+  - The default transformation in `FieldCoproductHint` changed from converting class names to lower case to converting
+    them to kebab case (e.g. `ConfigOption` is converted to `config-option`). The old behavior can be retained by
+    putting in scope an `implicit def coproductHint[T] = new FieldCoproductHint[T]("type") { override def
+    fieldValue(name: String): String = name.toLowerCase }`.
+
+- New features
+  - Added `deriveEnumerationReader`, `deriveEnumerationWriter` and `deriveEnumerationConvert` to the
+    `pureconfig.generic.semiauto` package, allowing the derivation of readers and writers for enumerations encoded as
+    sealed traits of case objects. As a consequence, the `EnumCoproductHint` is now deprecated in favor of these new
+    methods;
+  - `CoproductHint` now exposes a `noOptionFound` method allowing implementations to configure the failure reasons given
+    when no option for a sealed family was able to be read.
+
 ### 0.10.2 (Feb 5, 2019)
 
 - New features